Crate megazord

Re-exports

pub use fxaclient_ffi;

pub use logins_ffi;

pub use places_ffi;

pub use push_ffi;

pub use rc_log_ffi;

pub use sync_manager_ffi;

pub use viaduct;

Structs

StaticCStringPtr

VERSION_PTR

Statics

VERSION

Functions

full_megazord_get_version
In order to support the use case of consumers who don't know about megazords and don't need our e.g. networking or logging, we consider initialization optional for the default (full) megazord.